+++ Changelog.md
@@ -6,10 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased] Red Core Version 0.3 Changelog
 
+### Added
+
+- **RedClientTicker:** A useful ticker that allows you to have ticks every 2, 5 or 10 normal ticks, for things that shouldn't run 20 times a second.
+- **OptiNotFine:** A must-have when working with OptiFine compatibility, it allows you to know if OptiFine is installed, whether shaders are loaded, and to force fast render off.
+
 ### Changed
 
 - RedLogger now uses a String instead of a URI for the `newIssueLink` parameter, the old constructor has been deprecated and is due for removal for 0.4
 - RedLogger recomfort messages might be more random
+
 ### Optimized
 
 - Optimized `RedLogger#printFramedError` it should now use slightly less memory and thus have less GC pressure